Chocolate Randomiser
I'm not sure how it works but it does. After buying the factory, profit or MIB Alert appears on the 1st of every month at 0:30. And after the MIB decision it doesn't appear. This results in a fully functioning factory of monthly profits that has a chance to be shut down and is different every playthrough.