Chocolate events (#73)
* Chocolate Factory Event

Fun little event. To do:
1) Make STR_CHOCOLATE_FACTORY_PROFITS proc every month.
2) Balance the amount of funds and loyalty gained from each option.

* Update eventScripts_FTA.rul

* Chocolate Randomiser

I'm not sure how it works but it does. After buying the factory, profit or MIB Alert appears on the 1st of every month at 0:30. And after the MIB decision it doesn't appear. This results in a fully functioning factory of monthly profits that has a chance to be shut down and is different every playthrough.

* adjustments

Co-authored-by: Finnik <45336117+Finnik723@users.noreply.github.com>